[FIX] Fix code/reference mask handling
