+++ product_code_builder/models/__init__.py
@@ -17,5 +17,4 @@
 
 from . import product_attribute
 from . import product_attribute_value
-from . import product_product
-from . import product_template
+from . import product
